Keepin ModelPage as ModelPage to reduce scope
