+++ src/components/LocationPage/ChartsHolder.style.tsx
@@ -1,18 +1,13 @@
 import styled from 'styled-components';
 
-export const MainContentWrapper = styled.div`
+export const ChartContentWrapper = styled.div`
   margin-top: 85px;
-
   @media (min-width: 1350px) {
     margin-top: 97px;
   }
 `;
 
 export const MainContentInner = styled.div`
-  margin: 0;
-`;
-
-export const MainContentInnerBody = styled.div`
   margin: 0;
 
   @media (min-width: 900px) {
